Feat(Item Skill): added skill for weapon and shield
