+++ Assets/Scripts/Inventory/New Inventory/Scriptable Objects/Items/Scripts/ItemObject.cs
@@ -32,6 +32,7 @@ public class ItemObject : ScriptableObject
     [TextArea(15, 20)]
     public string description;
     public Item data = new Item();
+     public ScriptableObject skillData;
     
     public List<string> boneNames = new List<string>();
